requirements.txt
# ========================================
# JARVIS AI ASSISTANT - DEPENDENCIES
# Zero-Cost / High Performance Stack
# ========================================

# --- Core ---
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# --- Brain (LLM) ---
google-generativeai>=0.3.0
groq>=0.4.0

# --- Ears (Speech-to-Text) ---
faster-whisper>=0.10.0
sounddevice>=0.4.6
numpy>=1.24.0
# Note: Using numpy-based VAD instead of webrtcvad (avoids C++ build requirement)

# --- Mouth (Text-to-Speech) ---
edge-tts>=6.1.9
pygame>=2.5.0

# --- Memory (Vector Store) ---
chromadb>=0.4.22
sentence-transformers>=2.2.2

# --- Vision ---
mediapipe>=0.10.9
opencv-python>=4.8.0
Pillow>=10.0.0
pyautogui>=0.9.54
mss>=9.0.1

# --- Tools & Automation ---
pyautogui>=0.9.54
mss>=9.0.1
requests>=2.31.0
aiohttp>=3.9.0
aiofiles>=23.2.1

# --- Utilities ---
rich>=13.7.0
colorama>=0.4.6
